requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
discord==1.0.1
discord.py==1.5.1
Flask==1.1.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
oauth2client==4.1.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.5.1
